backend/requirements.txt
fastapi>=0.115,<0.116
uvicorn[standard]>=0.32,<0.33
pydantic>=2.9,<3
openai>=1.50,<2
python-dotenv>=1.0,<2
